feat: add BASE_IMAGEs to image cache
