+++ dev/io.openliberty.org.testcontainers/build.gradle
@@ -31,6 +31,17 @@ task assembleTestContainerData(type: JavaExec) {
     classpath = configurations.runtime.plus(sourceSets.main.runtimeClasspath)
     main = "io.openliberty.org.testcontainers.generate.CacheFiles"
     args project.getProjectDir()
+    
+    // Debugging properties for testcontainers
+    systemProperties = [
+      'org.slf4j.simpleLogger.logFile': new File(project.getProjectDir(), 'build/CacheFiles.log'),
+      'org.slf4j.simpleLogger.showDateTime': 'true',
+      'org.slf4j.simpleLogger.dateTimeFormat': '[MM/dd/yyyy HH:mm:ss:SSS z]',
+      'org.slf4j.simpleLogger.log.org.testcontainers': 'debug',
+      'org.slf4j.simpleLogger.log.tc': 'debug',
+      'org.slf4j.simpleLogger.log.com.github.dockerjava': 'warn',
+      'org.slf4j.simpleLogger.log.com.github.dockerjava.zerodep.shaded.org.apache.hc.client5.http.wire': 'off'
+    ]
 }
 
 task generateCustomImages(type: JavaExec) {
@@ -51,7 +62,7 @@ task generateCustomImages(type: JavaExec) {
     
     // Debugging properties for testcontainers
     systemProperties = [
-      'org.slf4j.simpleLogger.logFile': new File(project.getProjectDir(), 'build/testcontainers.log'),
+      'org.slf4j.simpleLogger.logFile': new File(project.getProjectDir(), 'build/CustomImages.log'),
       'org.slf4j.simpleLogger.showDateTime': 'true',
       'org.slf4j.simpleLogger.dateTimeFormat': '[MM/dd/yyyy HH:mm:ss:SSS z]',
       'org.slf4j.simpleLogger.log.org.testcontainers': 'debug',
