+++ dev/io.openliberty.org.testcontainers/src/io/openliberty/org/testcontainers/generate/CacheFiles.java
@@ -16,6 +16,8 @@ import java.io.BufferedWriter;
 import java.io.File;
 import java.nio.charset.Charset;
 import java.nio.file.Files;
+import java.nio.file.Path;
+import java.nio.file.Paths;
 import java.util.Arrays;
 import java.util.HashSet;
 import java.util.List;
@@ -109,7 +111,13 @@ public class CacheFiles {
             }
         }
         
-        //TODO Add all BASE_IMAGE arguments from Dockerfiles in this project
+        // Investigate all Dockerfiles and add the BASE_NAME to externals list
+		Path commonPath = Paths.get(projectPath, "resources", "openliberty", "testcontainers");
+		Dockerfile.findDockerfiles(commonPath).stream()
+			.map(location -> new Dockerfile(location))
+			.forEach(dockerfile -> {
+				externals.add(dockerfile.baseImageName.asCanonicalNameString());
+			});
         
         String header = "# NOTICE: This file was automatically updated to reflect changes made to test projects." + System.lineSeparator() +  
                         "# Please check these changes into GitHub" + System.lineSeparator();
